+++ pcap4j-core/src/main/java/org/pcap4j/packet/IpV6ExtRoutingPacket.java
@@ -15,6 +15,7 @@ import org.pcap4j.packet.factory.PacketFactories;
 import org.pcap4j.packet.factory.PacketFactory;
 import org.pcap4j.packet.namednumber.IpNumber;
 import org.pcap4j.packet.namednumber.IpV6RoutingHeaderType;
+import org.pcap4j.packet.namednumber.NotApplicable;
 import org.pcap4j.util.ByteArrays;
 
 /**
@@ -60,10 +61,15 @@ public final class IpV6ExtRoutingPacket extends AbstractPacket {
       Class<? extends Packet> nextPacketClass = factory.getTargetClass(header.getNextHeader());
       Packet nextPacket;
       if (nextPacketClass.equals(factory.getTargetClass())) {
-        try {
-          nextPacket
-            = IpV6ExtUnknownPacket.newPacket(rawData, offset + header.length(), payloadLength);
-        } catch (IllegalRawDataException e) {
+        nextPacket =
+          PacketFactories.getFactory(Packet.class, NotApplicable.class)
+            .newInstance(
+               rawData,
+               offset + header.length(),
+               payloadLength,
+               NotApplicable.UNKNOWN_IP_V6_EXTENSION
+             );
+        if (nextPacket instanceof IllegalPacket) {
           nextPacket = factory.newInstance(rawData, offset + header.length(), payloadLength);
         }
       }