+++ pcap4j-core/src/main/java/org/pcap4j/packet/IpV6NeighborDiscoveryRedirectedHeaderOption.java
@@ -13,7 +13,7 @@ import org.pcap4j.packet.IcmpV6CommonPacket.IpV6NeighborDiscoveryOption;
 import org.pcap4j.packet.factory.PacketFactories;
 import org.pcap4j.packet.namednumber.EtherType;
 import org.pcap4j.packet.namednumber.IpV6NeighborDiscoveryOptionType;
-import org.pcap4j.packet.namednumber.NA;
+import org.pcap4j.packet.namednumber.NotApplicable;
 import org.pcap4j.util.ByteArrays;
 
 /**
@@ -150,7 +150,7 @@ implements IpV6NeighborDiscoveryOption {
       byte[] ipRawData = p.get(IllegalPacket.class).getRawData();
       builder.getOuterOf(IllegalPacket.Builder.class)
         .payloadBuilder(
-          PacketFactories.getFactory(Packet.class, NA.class)
+          PacketFactories.getFactory(Packet.class, NotApplicable.class)
             .newInstance(ipRawData, 0, ipRawData.length)
               .getBuilder()
          );