+++ pcap4j-packetfactory-propertiesbased/src/main/java/org/pcap4j/packet/factory/PropertiesBasedIpV6TrafficClassFactory.java
@@ -10,7 +10,7 @@ package org.pcap4j.packet.factory;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import org.pcap4j.packet.IpV6Packet.IpV6TrafficClass;
-import org.pcap4j.packet.namednumber.NA;
+import org.pcap4j.packet.namednumber.NotApplicable;
 import org.pcap4j.util.ByteArrays;
 
 /**
@@ -18,7 +18,7 @@ import org.pcap4j.util.ByteArrays;
  * @since pcap4j 0.9.14
  */
 public final class PropertiesBasedIpV6TrafficClassFactory
-implements PacketFactory<IpV6TrafficClass, NA> {
+implements PacketFactory<IpV6TrafficClass, NotApplicable> {
 
   private static final PropertiesBasedIpV6TrafficClassFactory INSTANCE
     = new PropertiesBasedIpV6TrafficClassFactory();
@@ -33,7 +33,7 @@ implements PacketFactory<IpV6TrafficClass, NA> {
 
   @Override
   @Deprecated
-  public IpV6TrafficClass newInstance(byte[] rawData, int offset, int length, NA number) {
+  public IpV6TrafficClass newInstance(byte[] rawData, int offset, int length, NotApplicable number) {
     return newInstance(rawData, offset, length);
   }
 
@@ -82,7 +82,7 @@ implements PacketFactory<IpV6TrafficClass, NA> {
 
   @Override
   @Deprecated
-  public Class<? extends IpV6TrafficClass> getTargetClass(NA number) {
+  public Class<? extends IpV6TrafficClass> getTargetClass(NotApplicable number) {
     return getTargetClass();
   }