+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/PacketFactoryBinder.java
@@ -24,7 +24,7 @@ import org.pcap4j.packet.namednumber.EtherType;
 import org.pcap4j.packet.namednumber.IcmpV4Type;
 import org.pcap4j.packet.namednumber.IcmpV6Type;
 import org.pcap4j.packet.namednumber.IpNumber;
-import org.pcap4j.packet.namednumber.NA;
+import org.pcap4j.packet.namednumber.NotApplicable;
 import org.pcap4j.packet.namednumber.NamedNumber;
 import org.pcap4j.packet.namednumber.TcpPort;
 import org.pcap4j.packet.namednumber.UdpPort;
@@ -50,7 +50,7 @@ final class PacketFactoryBinder {
     packetFactories.put(IpNumber.class, StaticIpNumberPacketFactory.getInstance());
     packetFactories.put(TcpPort.class, StaticTcpPortPacketFactory.getInstance());
     packetFactories.put(UdpPort.class, StaticUdpPortPacketFactory.getInstance());
-    packetFactories.put(NA.class, StaticUnknownPacketFactory.getInstance());
+    packetFactories.put(NotApplicable.class, StaticNotApplicablePacketFactory.getInstance());
 
     packetpPieceFactories.put(
       IpV4Option.class,