+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/StaticIpV6FlowLabelFactory.java
@@ -10,7 +10,7 @@ package org.pcap4j.packet.factory;
 import static org.pcap4j.util.ByteArrays.*;
 import org.pcap4j.packet.IpV6Packet.IpV6FlowLabel;
 import org.pcap4j.packet.IpV6SimpleFlowLabel;
-import org.pcap4j.packet.namednumber.NA;
+import org.pcap4j.packet.namednumber.NotApplicable;
 import org.pcap4j.util.ByteArrays;
 
 /**
@@ -18,7 +18,7 @@ import org.pcap4j.util.ByteArrays;
  * @since pcap4j 0.9.16
  */
 public final class StaticIpV6FlowLabelFactory
-implements PacketFactory<IpV6FlowLabel, NA> {
+implements PacketFactory<IpV6FlowLabel, NotApplicable> {
 
   private static final StaticIpV6FlowLabelFactory INSTANCE
     = new StaticIpV6FlowLabelFactory();
@@ -33,7 +33,7 @@ implements PacketFactory<IpV6FlowLabel, NA> {
 
   @Override
   @Deprecated
-  public IpV6FlowLabel newInstance(byte[] rawData, int offset, int length, NA number) {
+  public IpV6FlowLabel newInstance(byte[] rawData, int offset, int length, NotApplicable number) {
     return newInstance(rawData, offset, length);
   }
 
@@ -56,7 +56,7 @@ implements PacketFactory<IpV6FlowLabel, NA> {
 
   @Override
   @Deprecated
-  public Class<? extends IpV6FlowLabel> getTargetClass(NA number) {
+  public Class<? extends IpV6FlowLabel> getTargetClass(NotApplicable number) {
     return getTargetClass();
   }